+++ examples/blockfrost/index.js
@@ -38,13 +38,12 @@ let address = Core.addressFromBech32(
 );
 
 const wallet = new ColdWallet(address, 0, provider);
-const blaze = new Blaze(provider, wallet);
+const blaze = await Blaze.from(provider, wallet);
 
 console.log(await provider.getParameters());
 
-const tx = await (
-  await blaze.newTransaction()
-)
+const tx = await blaze
+  .newTransaction()
   .payLovelace(
     Address.fromBech32(
       // Some other preview address with BTN
@@ -71,10 +70,10 @@ for (const utxo of utxos) {
   console.log(utxoRef);
 
   const amountADA = utxo.output().amount().coin();
-  console.log(`Amount of ADA: ${amountADA / 1000000}`);
+  console.log(`Amount of ADA: ${amountADA / 1000000n}`);
 
   const amountBTN = utxo.output().amount().multiasset().get(btnUnit);
-  console.log(`Amount of ${assetName}: ${amountBTN / 1000000}`);
+  console.log(`Amount of ${assetName}: ${amountBTN / 1000000n}`);
 }
 
 // Some NFT on Preview
@@ -93,13 +92,21 @@ const txIns = [
     "74baf638a18a6ab54798cac310112af61cb1f1d6eacb8c893a10b6877cf71a8d",
     1,
   ),
+  new TransactionInput(
+    "036ed48c89169a9c4e475e08a6d22ea1e09ba8a6a6cfbabfa4f1deefd269652c",
+    0,
+  ),
+  new TransactionInput(
+    "ad54aa407df81404ab74343a4fca56e51c4ab488c54f12c4bc43e8b093c976a5",
+    1,
+  ),
 ];
 
 const resolvedOutputs = await provider.resolveUnspentOutputs(txIns);
 
 for (const utxo of resolvedOutputs) {
   const amountADA = utxo.output().amount().coin();
-  console.log(`Amount of ADA: ${amountADA / 1000000}`);
+  console.log(`Amount of ADA: ${amountADA / 1000000n}`);
 
   const multiAssetMap = utxo.output().amount().multiasset();
   for (const [asset, amount] of multiAssetMap.entries()) {
@@ -110,6 +117,16 @@ for (const utxo of resolvedOutputs) {
     ).toString();
 
     console.log(`Amount of ${assetName}: ${amount}`);
+  }
+
+  if (utxo.output().datum()) {
+    console.log("Datum:");
+    console.log(utxo.output().datum().toCore());
+  }
+
+  if (utxo.output().scriptRef()) {
+    console.log("Reference Script:");
+    console.log(utxo.output().scriptRef().hash());
   }
 }