+++ packages/blaze-blueprint/test/blueprint_generate.ts
@@ -1,0 +1,5 @@
+import { generateBlueprint } from "../src";
+
+generateBlueprint({
+  infile: "contracts/plutus.json",
+});