+++ packages/blaze-sdk/src/blaze.ts
@@ -1,4 +1,8 @@
-import type { Transaction } from "@blaze-cardano/core";
+import type {
+  ProtocolParameters,
+  Transaction,
+  TransactionId,
+} from "@blaze-cardano/core";
 import { CborSet, VkeyWitness } from "@blaze-cardano/core";
 import type { Provider } from "@blaze-cardano/query";
 import { TxBuilder } from "@blaze-cardano/tx";
@@ -9,32 +13,47 @@ import type { Wallet } from "@blaze-cardano/wallet";
  * It requires a provider and a wallet to interact with the blockchain and manage funds.
  */
 export class Blaze<ProviderType extends Provider, WalletType extends Wallet> {
-  provider: ProviderType;
+  readonly provider: ProviderType;
   wallet: WalletType;
+  readonly params: ProtocolParameters;
 
   /**
-   * Constructs a new instance of the blaze class.
+   * Constructs a new instance of the Blaze class.
    * @param {ProviderType} provider - The provider to use for interacting with the blockchain.
-   * @param {WalletType} wallet - The wallet to use for managing funds.
+   * @param {WalletType} wallet - The wallet to use for managing funds and signing transactions.
+   * @private
    */
-  constructor(provider: ProviderType, wallet: WalletType) {
+  private constructor(
+    provider: ProviderType,
+    wallet: WalletType,
+    params: ProtocolParameters,
+  ) {
     this.provider = provider;
     this.wallet = wallet;
+    this.params = params;
+  }
+
+  static async from<ProviderType extends Provider, WalletType extends Wallet>(
+    provider: ProviderType,
+    wallet: WalletType,
+  ) {
+    const params = await provider.getParameters();
+    return new Blaze(provider, wallet, params);
   }
 
   /**
    * Creates a new transaction using the provider and wallet.
    * @returns {TxBuilder} - The newly created transaction builder.
    */
-  async newTransaction() {
-    const params = await this.provider.getParameters();
-    const myUtxos = await this.wallet.getUnspentOutputs();
-    const changeAddress = await this.wallet.getChangeAddress();
-    return new TxBuilder(params)
-      .setNetworkId(await this.wallet.getNetworkId())
-      .addUnspentOutputs(myUtxos)
-      .setChangeAddress(changeAddress)
-      .useEvaluator((x, y) => this.provider.evaluateTransaction(x, y));
+  newTransaction() {
+    return new TxBuilder(this.params).addPreCompleteHook(async (tx) => {
+      const myUtxos = await this.wallet.getUnspentOutputs();
+      const changeAddress = await this.wallet.getChangeAddress();
+      tx.setNetworkId(await this.wallet.getNetworkId())
+        .addUnspentOutputs(myUtxos)
+        .setChangeAddress(changeAddress)
+        .useEvaluator((x, y) => this.provider.evaluateTransaction(x, y));
+    });
   }
 
   /**
@@ -68,5 +87,17 @@ export class Blaze<ProviderType extends Provider, WalletType extends Wallet> {
     );
     tx.setWitnessSet(ws);
     return tx;
+  }
+
+  /**
+   * Submits a transaction to the blockchain.
+   * @param {Transaction} tx - The transaction to submit.
+   * @returns {Promise<TransactionId>} - The transaction ID.
+   * @throws {Error} If the transaction submission fails.
+   * @description This method sends the provided transaction to the blockchain network
+   * using the configured provider.
+   */
+  async submitTransaction(tx: Transaction): Promise<TransactionId> {
+    return this.provider.postTransactionToChain(tx);
   }
 }