+++ packages/blaze-uplc/src/index.ts
@@ -1,2 +1,2 @@
 /* Unfortunately no UPLC machine yet */
-export { applyParams } from "./utils";
+export * from "./utils";